tested auth apis and fixed numerous bugs

+++ authentication/views.py
@@ -1,5 +1,6 @@
 from django.http.response import JsonResponse
 from django.shortcuts import render
+from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
 from rest_framework import serializers, status
 from rest_framework import generics
@@ -23,19 +24,21 @@ def register_admin(request):
     serialized = UserSerializer(data = request.data)
     if(not Administrator.objects.exists()):
         if serialized.is_valid():
-            user = User.objects.create(
-                username = request.data.get('username'),
-                email = request.data.get('email'),
-                first_name = request.data.get('first_name'),
-                last_name = request.data.get('last_name'),
-                user_type = request.data.get('user_type'),
-                # password = make_password(request.data.get('password')) + "_NEW",
-            )
-            user.set_password(request.data.get('password'))
-            user.save()
-            administrator = Administrator.objects.create(user = user)
-            print(user.password)
-            return Response(serialized.data, status=status.HTTP_201_CREATED)
+            if request.data.get("user_type") == 1:
+                user = User.objects.create(
+                    username = request.data.get('username'),
+                    email = request.data.get('email'),
+                    first_name = request.data.get('first_name'),
+                    last_name = request.data.get('last_name'),
+                    user_type = request.data.get('user_type'),
+                    password = make_password(request.data.get('password')) + "_NEW",
+                )
+                # user.set_password(request.data.get('password'))
+                # user.save()
+                administrator = Administrator.objects.create(user = user)
+                return Response(serialized.data, status=status.HTTP_201_CREATED)
+            else:
+                return Response("Can only register an admin")
         else:
             return Response(serialized._errors, status=status.HTTP_400_BAD_REQUEST)
     else:
@@ -46,11 +49,13 @@ def register_admin(request):
 @permission_classes([AdministratorPermission])
 def delete_admin(request):
     try:
-        print(User.objects.get(username=request.data.get('username')))
         user = User.objects.get(username=request.data.get('username'))
-        admin = Administrator.objects.get(user=user)
-        admin.delete()
-        user.delete()
+        if user.user_type == 1:
+            admin = Administrator.objects.get(user=user)
+            admin.delete()
+            user.delete()
+        else:
+            return Response({"error": "Can only delete a teacher"})
     except:
         return Response({"error": "The user does not exist"})
     else:
@@ -67,11 +72,12 @@ def login_user(request):
     try:
         user = User.objects.get(username=request.data.get('username'))
         if(user.password[-4:] == "_NEW"):
-            print('haha gotcha')
             return Response("Go change yo password")
         else:
-            print(user.check_password(request.data.get('password')))
-            return Response(get_tokens_for_user(user))
+            if user.check_password(request.data.get('password')):
+                return Response(get_tokens_for_user(user))
+            else:
+                return Response("Wrong password")
     except:
         return Response("OOPS")
 
@@ -118,21 +124,25 @@ class ChangePasswordView(generics.UpdateAPIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['POST'])
+@permission_classes([AdministratorPermission])
 def register_student(request):
     serialized = UserSerializer(data = request.data)
     if serialized.is_valid():
-        user = User.objects.create(
-            username = request.data.get('username'),
-            email = request.data.get('email'),
-            first_name = request.data.get('first_name'),
-            last_name = request.data.get('last_name'),
-            user_type = request.data.get('user_type'),
-            # password = make_password(request.data.get('password')) + "_NEW",
-        )
-        user.set_password(request.data.get('password'))
-        user.save()
-        student = Student.objects.create(user = user)
-        return Response(serialized.data, status=status.HTTP_201_CREATED)
+        if request.data.get("user_type") == 3:
+            user = User.objects.create(
+                username = request.data.get('username'),
+                email = request.data.get('email'),
+                first_name = request.data.get('first_name'),
+                last_name = request.data.get('last_name'),
+                user_type = request.data.get('user_type'),
+                # password = make_password(request.data.get('password')) + "_NEW",
+            )
+            user.set_password(request.data.get('password'))
+            user.save()
+            student = Student.objects.create(user = user)
+            return Response(serialized.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response("Can only register a student")
     else:
         return Response(serialized._errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -141,9 +151,12 @@ def register_student(request):
 def delete_student(request):
     try:
         user = User.objects.get(username=request.data.get('username'))
-        student = Student.objects.get(user=user)
-        student.delete()
-        user.delete()
+        if user.user_type == 3:
+            student = Student.objects.get(user=user)
+            student.delete()
+            user.delete()
+        else:
+            return Response({"error": "Can only delete a teacher"})
     except:
         return Response({"error": "The user does not exist"})
     else:
@@ -156,21 +169,25 @@ def get_student(request):
     return JsonResponse({"data": {"student": student, "user": user}})
 
 @api_view(['POST'])
+@permission_classes([AdministratorPermission])
 def register_teacher(request):
     serialized = UserSerializer(data = request.data)
     if serialized.is_valid():
-        user = User.objects.create(
-            username = request.data.get('username'),
-            email = request.data.get('email'),
-            first_name = request.data.get('first_name'),
-            last_name = request.data.get('last_name'),
-            user_type = request.data.get('user_type'),
-            # password = make_password(request.data.get('password')) + "_NEW",
-        )
-        user.set_password(request.data.get('password'))
-        user.save()
-        teacher = Teacher.objects.create(user = user)
-        return Response(serialized.data, status=status.HTTP_201_CREATED)
+        if request.data.get("user_type" == 2):
+            user = User.objects.create(
+                username = request.data.get('username'),
+                email = request.data.get('email'),
+                first_name = request.data.get('first_name'),
+                last_name = request.data.get('last_name'),
+                user_type = request.data.get('user_type'),
+                # password = make_password(request.data.get('password')) + "_NEW",
+            )
+            user.set_password(request.data.get('password'))
+            user.save()
+            teacher = Teacher.objects.create(user = user)
+            return Response(serialized.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response("Can only register a teacher")
     else:
         return Response(serialized._errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -179,13 +196,16 @@ def register_teacher(request):
 def delete_teacher(request):
     try:
         user = User.objects.get(username=request.data.get('username'))
-        teacher = Teacher.objects.get(user=user)
-        teacher.delete()
-        user.delete()
+        if user.user_type == 2:
+            teacher = Teacher.objects.get(user=user)
+            teacher.delete()
+            user.delete()
+        else:
+            return Response({"error": "Can only delete a teacher"})
     except:
-        return Response({"error": "The user does not exist"})
+        return Response({"error": "The teacher does not exist"})
     else:
-        return Response({"success": "The user was deleted"})
+        return Response({"success": "The teacher was deleted"})
 
 @api_view(['GET'])
 def get_teacher(request):
